+++ app/controllers/api/encryptables_controller.rb
@@ -19,6 +19,8 @@ class Api::EncryptablesController < ApiController
   # GET /api/encryptables/:id
   def show
     authorize encryptable
+    encryptable.paper_trail_event = 'viewed'
+    encryptable.touch
     encryptable.decrypt(decrypted_team_password(team))
     render_entry
   end